# specs/2-auth-jwt/contracts/openapi.yaml
openapi: 3.0.3
info:
  title: Authentication API - Auth-JWT-Bridge
  description: API for authentication and JWT token management in the multi-user todo application
  version: 1.0.0
servers:
  - url: http://localhost:8000
    description: Development server
  - url: https://api.example.com
    description: Production server
paths:
  /api/auth/login:
    post:
      summary: User login
      description: Authenticate user with email and password, return JWT token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Successful authentication
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponse'
        '400':
          description: Invalid credentials or bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized - Invalid credentials
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/auth/signup:
    post:
      summary: User registration
      description: Register a new user with email and password, return JWT token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SignupRequest'
      responses:
        '200':
          description: User registered successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SignupResponse'
        '400':
          description: Invalid input data or email already exists
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/auth/logout:
    post:
      summary: User logout
      description: Invalidate user session
      responses:
        '200':
          description: Successfully logged out
        '401':
          description: Unauthorized - Invalid or missing token
  /api/{user_id}/tasks:
    get:
      summary: List user's tasks
      description: Retrieve a list of tasks for the authenticated user
      parameters:
        - name: user_id
          in: path
          required: true
          description: The ID of the user whose tasks to retrieve
          schema:
            type: string
        - name: status
          in: query
          required: false
          description: Filter tasks by completion status
          schema:
            type: string
            enum: [all, completed, incomplete]
            default: all
      responses:
        '200':
          description: List of tasks for the user
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Task'
        '401':
          description: Unauthorized - Invalid or missing JWT token
        '403':
          description: Forbidden - User ID in token doesn't match path parameter
        '404':
          description: User not found
      security:
        - bearerAuth: []
    post:
      summary: Create a new task
      description: Create a new task for the authenticated user
      parameters:
        - name: user_id
          in: path
          required: true
          description: The ID of the user creating the task
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateTaskRequest'
      responses:
        '201':
          description: Task created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Task'
        '400':
          description: Bad request - Invalid input data
        '401':
          description: Unauthorized - Invalid or missing JWT token
        '403':
          description: Forbidden - User ID in token doesn't match path parameter
        '404':
          description: User not found
      security:
        - bearerAuth: []
  /api/{user_id}/tasks/{id}:
    get:
      summary: Get a specific task
      description: Retrieve details of a specific task
      parameters:
        - name: user_id
          in: path
          required: true
          description: The ID of the user
          schema:
            type: string
        - name: id
          in: path
          required: true
          description: The ID of the task to retrieve
          schema:
            type: string
      responses:
        '200':
          description: Task details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Task'
        '401':
          description: Unauthorized - Invalid or missing JWT token
        '403':
          description: Forbidden - User ID in token doesn't match path parameter or task doesn't belong to user
        '404':
          description: Task not found
      security:
        - bearerAuth: []
    put:
      summary: Update a task
      description: Update details of an existing task
      parameters:
        - name: user_id
          in: path
          required: true
          description: The ID of the user
          schema:
            type: string
        - name: id
          in: path
          required: true
          description: The ID of the task to update
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateTaskRequest'
      responses:
        '200':
          description: Task updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Task'
        '400':
          description: Bad request - Invalid input data
        '401':
          description: Unauthorized - Invalid or missing JWT token
        '403':
          description: Forbidden - User ID in token doesn't match path parameter or task doesn't belong to user
        '404':
          description: Task not found
      security:
        - bearerAuth: []
    delete:
      summary: Delete a task
      description: Permanently delete a task
      parameters:
        - name: user_id
          in: path
          required: true
          description: The ID of the user
          schema:
            type: string
        - name: id
          in: path
          required: true
          description: The ID of the task to delete
          schema:
            type: string
      responses:
        '204':
          description: Task deleted successfully
        '401':
          description: Unauthorized - Invalid or missing JWT token
        '403':
          description: Forbidden - User ID in token doesn't match path parameter or task doesn't belong to user
        '404':
          description: Task not found
      security:
        - bearerAuth: []
  /api/{user_id}/tasks/{id}/complete:
    patch:
      summary: Toggle task completion status
      description: Toggle the completion status of a task
      parameters:
        - name: user_id
          in: path
          required: true
          description: The ID of the user
          schema:
            type: string
        - name: id
          in: path
          required: true
          description: The ID of the task to update
          schema:
            type: string
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                completed:
                  type: boolean
                  description: New completion status (optional, will toggle if not provided)
      responses:
        '200':
          description: Task completion status updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Task'
        '401':
          description: Unauthorized - Invalid or missing JWT token
        '403':
          description: Forbidden - User ID in token doesn't match path parameter or task doesn't belong to user
        '404':
          description: Task not found
      security:
        - bearerAuth: []

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    LoginRequest:
      type: object
      required:
        - email
        - password
      properties:
        email:
          type: string
          format: email
          description: User's email address
        password:
          type: string
          format: password
          description: User's password
    LoginResponse:
      type: object
      required:
        - token
        - expires_in
        - user
      properties:
        token:
          type: string
          description: JWT token for authentication
        expires_in:
          type: integer
          description: Number of seconds until token expiration
        user:
          $ref: '#/components/schemas/User'
    SignupRequest:
      type: object
      required:
        - email
        - password
      properties:
        email:
          type: string
          format: email
          description: User's email address
        password:
          type: string
          format: password
          description: User's password
        name:
          type: string
          description: User's display name (optional)
    SignupResponse:
      type: object
      required:
        - token
        - expires_in
        - user
      properties:
        token:
          type: string
          description: JWT token for authentication
        expires_in:
          type: integer
          description: Number of seconds until token expiration
        user:
          $ref: '#/components/schemas/User'
    User:
      type: object
      required:
        - id
        - email
      properties:
        id:
          type: string
          format: uuid
          description: Unique identifier for the user
        email:
          type: string
          format: email
          description: User's email address
        name:
          type: string
          description: User's display name (optional)
        created_at:
          type: string
          format: date-time
          description: Timestamp when the user was created
    Task:
      type: object
      required:
        - id
        - title
        - completed
        - user_id
        - created_at
        - updated_at
      properties:
        id:
          type: string
          format: uuid
          description: Unique identifier for the task
        title:
          type: string
          minLength: 1
          maxLength: 200
          description: Task title (1-200 characters)
        description:
          type: string
          nullable: true
          description: Optional task description
        completed:
          type: boolean
          description: Whether the task is completed
        user_id:
          type: string
          format: uuid
          description: ID of the user who owns this task
        created_at:
          type: string
          format: date-time
          description: Timestamp when the task was created
        updated_at:
          type: string
          format: date-time
          description: Timestamp when the task was last updated
    CreateTaskRequest:
      type: object
      required:
        - title
      properties:
        title:
          type: string
          minLength: 1
          maxLength: 200
          description: Task title (1-200 characters)
        description:
          type: string
          nullable: true
          description: Optional task description
    UpdateTaskRequest:
      type: object
      properties:
        title:
          type: string
          minLength: 1
          maxLength: 200
          description: Updated task title (1-200 characters)
        description:
          type: string
          nullable: true
          description: Updated task description
        completed:
          type: boolean
          description: Updated completion status
    ErrorResponse:
      type: object
      required:
        - error
        - message
      properties:
        error:
          type: string
          description: Error code
        message:
          type: string
          description: Human-readable error message
        status_code:
          type: integer
          description: HTTP status code